Fixed the exit_date issue with 2009-2010. Because some dates were converted to strings formatted like dates (i.e. 06/22/2010) rather than SPSS-origin seconds from the epoch (i.e. 13333485012), I had to selectively convert some dates.
The solution was to use grep to find any instances of punctuation, then use as.Date to convert those instances to an R Date, then use difftime to get the SPSS formatted date with the SPSS epoch (1582-10-14) as the begin time. This way, all of the dates can then be handled by the same reformat to R dates.

There were also several input errors in the disability category related to single-letter disability types (other than A which signifies Autism). I added this to the gsub regex to remove those and replace with a blank to be coded NA.